reconstruction/pyproject.toml
[project]
name = "reconstruction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.19"
dependencies = [
    "fused-bilagrid",
    "fused-ssim",
    "gsplat>=1.5.3",
    "imageio[ffmpeg]>=2.37.3",
    "matplotlib>=3.10.8",
    "nerfview>=0.1.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "pycolmap",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "splines>=0.3.3",
    "tensorboard>=2.20.0",
    "tensorly>=0.9.0",
    "torch>=2.11.0",
    "torchmetrics[image]>=1.9.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "tyro>=0.8.8",
    "viser>=1.0.26",
]

[tool.uv.sources]
fused-ssim = { git = "https://github.com/rahul-goel/fused-ssim", rev = "a7c48d6dd7ac6dc39a7958c7c4452e0b10418f38" }
fused-bilagrid = { git = "https://github.com/harry7557558/fused-bilagrid", rev = "3bd774b644c3194225a5a0ddf46157092511ba3d" }
pycolmap = { path = "../colmap" }

[tool.uv.extra-build-dependencies]
fused-bilagrid = ["torch>=2.11.0"]
fused-ssim = ["torch>=2.11.0"]